move the metadata verification and update out of schema and into tools
