+++ pyproject.toml
@@ -22,6 +22,7 @@ black = "^24.4.2"
 
 [tool.ruff]
 line-length = 120
+fix = true
 
 [tool.isort]
 line_length = 120
