+++ tests/test_schemas.py
@@ -1,15 +1,8 @@
-import json
-
 import pytest
 from pydantic import ValidationError
 
-from pymetadataeditor.schemas import (
-    SchemaBaseModel,
-    SeriesDescription,
-    TimeSeriesMetadataSchema,
-    update_metadata,
-    validate_metadata,
-)
+from pymetadataeditor.schemas import SchemaBaseModel, TimeSeriesMetadataSchema
+from pymetadataeditor.tools import validate_metadata
 
 
 def test_SchemaBaseModel():
@@ -30,39 +23,6 @@ def test_SchemaBaseModel():
 
         # nor at instantiation
         SubClassTest(f1=1, f2="two", f3=3)
-
-
-def test_validate_metadata():
-    class Example(SchemaBaseModel):
-        idno: str
-        series_description: SeriesDescription
-
-    with pytest.raises(AssertionError):
-        validate_metadata({}, "bad_schema")
-
-    data_is_malformed_json_string = """{"idno": "GB123"""
-    with pytest.raises(json.decoder.JSONDecodeError):
-        validate_metadata(data_is_malformed_json_string, Example)
-
-    data_is_incomplete_string = """{"idno": "GB123"}"""
-    with pytest.raises(ValueError):
-        validate_metadata(data_is_incomplete_string, Example)
-
-    data_is_valid_string = """{"idno": "GB123", "series_description": {"idno": "GB123", "name": "GB123"}}"""
-    validate_metadata(data_is_valid_string, Example)
-
-    data_is_incomplete_dict = {"idno": "12"}
-    with pytest.raises(ValueError):
-        validate_metadata(data_is_incomplete_dict, Example)
-
-    data_is_valid_dict = {"idno": "12", "series_description": {"idno": "12", "name": "n"}}
-    validate_metadata(data_is_valid_dict, Example)
-
-    data_is_mix_of_dict_and_obj = {"idno": "12", "series_description": SeriesDescription(idno="12", name="n")}
-    validate_metadata(data_is_mix_of_dict_and_obj, Example)
-
-    data_is_obj = Example(idno="12", series_description=SeriesDescription(idno="12", name="n"))
-    validate_metadata(data_is_obj, Example)
 
 
 def test_validate_metadata_TimeSeries():
@@ -107,21 +67,3 @@ def test_validate_metadata_TimeSeries():
         },
     }
     validate_metadata(good_uri, TimeSeriesMetadataSchema)
-
-
-def test_update_metadata():
-    class Example(SchemaBaseModel):
-        f1: int
-        f2: str
-        f3: float
-
-    original = Example(f1=1, f2="two", f3=3.0)
-    new = update_metadata(original, f1=None, f3=3.3, f2="new_value")
-
-    assert original.f1 == 1
-    assert original.f2 == "two"
-    assert original.f3 == 3.0
-    assert isinstance(new, Example)
-    assert new.f1 == 1
-    assert new.f2 == "new_value"
-    assert new.f3 == 3.3
